fix: Fixed an issue with Ads in the MainActivity
When ad was loaded there was a issue with marginRight

MarginRight was set 2 times more than other margins. Now it is fixed and banner Ads does not overlap the content in MainActivity's RecyclerView. I also adjusted ads in tablet version of MainActivity, small change to dimension and slovak string files