fix: Fixed an UI issue with loading movies in MainActivity
noFavoriteMoviesTextView was shown after loading was done.

When user has no favorite movies in his collection, that textView is visible. When user changed section and loading was done, loaded movie posters covered most of that textView, but it was still visible. In this case there was no change to visibility in that textView after it's visibility was set to true. There is now a change which changes the visibility of that textView to INVISIBLE.